requirements.txt
numpy>=1.21.0
librosa>=0.9.0
torch>=1.9.0
torchaudio>=0.9.0
pyannote.audio>=2.1.1
scipy>=1.7.0
streamlit>=1.0.0
soundfile>=0.10.3
scikit-learn>=0.24.2
pydub>=0.25.1